Restructuring folders and files
